Se implementaron mejoras y se realizo la configuracion de PWA para permitir el uso como app movil o desktop
