+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "portfolio",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "private": true,
   "author": {
     "name": "Cristian Manco",
@@ -18,6 +18,7 @@
     "framer-motion": "^12.16.0",
     "lucide-react": "^0.512.0",
     "next": "^15.3.3",
+    "next-pwa": "^5.6.0",
     "nodemailer": "^7.0.3",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
@@ -26,6 +27,7 @@
     "react-type-animation": "^3.2.0"
   },
   "devDependencies": {
+    "@types/estree": "^1.0.8",
     "@types/node": "^22.15.29",
     "@types/nodemailer": "^6.4.17",
     "@types/react": "^19.1.6",
