+++ public/sw.js
@@ -1,0 +1,1 @@
+if(!self.define){let e,a={};const n=(n,s)=>(n=new URL(n+".js",s).href,a[n]||new Promise(a=>{if("document"in self){const e=document.createElement("script");e.src=n,e.onload=a,document.head.appendChild(e)}else e=n,importScripts(n),a()}).then(()=>{let e=a[n];if(!e)throw new Error(`Module ${n} didn’t register its module`);return e}));self.define=(s,i)=>{const c=e||("document"in self?document.currentScript.src:"")||location.href;if(a[c])return;let t={};const r=e=>n(e,c),o={module:{uri:c},exports:t,require:r};a[c]=Promise.all(s.map(e=>o[e]||r(e))).then(e=>(i(...e),t))}}define(["./workbox-4754cb34"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"5106ed6ce0a444dc6d39c9b1c353b71c"},{url:"/_next/static/IXMmktk8ekr6HIRaPny85/_buildManifest.js",revision:"acaa94a664c1fdcca9b72cec00ec7066"},{url:"/_next/static/IXMmktk8ekr6HIRaPny85/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/241-a66ff67ad31f2052.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/30a37ab2-cf110d9038233c7e.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/341.df3329d77a5faa19.js",revision:"df3329d77a5faa19"},{url:"/_next/static/chunks/353-97d34345cf393900.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/472.a3826d29d6854395.js",revision:"a3826d29d6854395"},{url:"/_next/static/chunks/4bd1b696-d1354b491c8c2d32.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/565-25c8cfd0c467cdc6.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/647-eb279da69bf22947.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/684-ea651bec21ccde65.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/689-ceb4f385b40fb629.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/699-d25f5f16389c526f.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/766-0b239707da818c86.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/874-af714c0684d68372.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/895-c28dfb153d08a234.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/8e1d74a4-6709ea36eadc11f6.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/app/_not-found/page-41eecf924ca0228d.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/app/about/page-21339410bb9fbda9.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/app/api/email/route-d08b0fe940529dc8.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/app/contact/page-3f7f0791d46c683b.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/app/layout-4b41c87bf1013007.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/app/not-found-6415327b3f7f1553.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/app/page-6fc32a1be2ca4373.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/app/portfolio/page-1053682b8f850aa5.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/app/skills/page-84d6dffee43918e2.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/framework-f593a28cde54158e.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/main-app-8bdf846b44d792aa.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/main-cce17225b9a0fac4.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/pages/_app-92f2aae776f86b9c.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/pages/_error-1c75af675ede1ae6.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-ff07440ecc80fdc5.js",revision:"IXMmktk8ekr6HIRaPny85"},{url:"/_next/static/css/a4ee324bb573b882.css",revision:"a4ee324bb573b882"},{url:"/_next/static/media/26a46d62cd723877-s.woff2",revision:"befd9c0fdfa3d8a645d5f95717ed6420"},{url:"/_next/static/media/55c55f0601d81cf3-s.woff2",revision:"43828e14271c77b87e3ed582dbff9f74"},{url:"/_next/static/media/581909926a08bbc8-s.woff2",revision:"f0b86e7c24f455280b8df606b89af891"},{url:"/_next/static/media/8e9860b6e62d6359-s.woff2",revision:"01ba6c2a184b8cba08b0d57167664d75"},{url:"/_next/static/media/97e0cb1ae144a2a9-s.woff2",revision:"e360c61c5bd8d90639fd4503c829c2dc"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/_next/static/media/e4af272ccee01ff0-s.p.woff2",revision:"65850a373e258f1c897a2b3d75eb74de"},{url:"/icons/icon-16x16.png",revision:"fad31c8e55634651e57fb7e2bba70cb5"},{url:"/icons/icon-192x192.png",revision:"da21a6d3b1f669708cde3a2bd4de660c"},{url:"/icons/icon-32x32.png",revision:"1064236ff71b2256fb8a987cfd3826de"},{url:"/icons/icon-512x512.png",revision:"041f8d940bba3a0217c77ab84063b9fe"},{url:"/img/Captura de pantalla 2024-08-30 214412.png",revision:"7835d23ae989ad9b496b62bb3cee6ce6"},{url:"/img/HV-CRISTIAN MANCO.pdf",revision:"52b43b88af0a7a65450a15730f6a0bcd"},{url:"/img/chat.png",revision:"b77f0838607592c85ec4b39e31408771"},{url:"/img/codigo.jpg",revision:"bedc0aff5677a2bdffc1003bb0b1873f"},{url:"/img/contact_bg.png",revision:"a3dff246d77211de05dd6a598ee0b094"},{url:"/img/develoment.png",revision:"83d00835b9f2ec7f071d56df406f3167"},{url:"/img/favicon.ico",revision:"c30c7d42707a47a3f4591831641e50dc"},{url:"/img/fondo.jpg",revision:"775737833215d97fe79e7412ca442b32"},{url:"/img/fondo.png",revision:"3c6467278a66af3216348e8588a69062"},{url:"/img/hola.png",revision:"83adbf99eaf56120cf45b9dadd213f01"},{url:"/img/image copy 2.png",revision:"584415e1066257ea5ad997cd27c96250"},{url:"/img/image copy 3.png",revision:"e533cdcc65c44a92aafb6dc6f3be2842"},{url:"/img/image copy 5.png",revision:"8553b6da7b43b645310936a2fdabec1e"},{url:"/img/image copy 6.png",revision:"9d53ea46b609c8efa3787ff722d5fca7"},{url:"/img/image copy 7.png",revision:"049a4d3b47b6c295d2e2b2db1d0fe43d"},{url:"/img/image copy 8.png",revision:"942e10074917df0d2a68929cb0b352af"},{url:"/img/image copy.png",revision:"97a01a47fe14d8f84cab38e9c1a99741"},{url:"/img/image.png",revision:"70c061858d0bb6adfb0c5e927be4ac6b"},{url:"/img/image1.png",revision:"aca950509d184f310debba78c2daa9fa"},{url:"/img/image2.png",revision:"a6b4bd46a48626135629d9cd4e208922"},{url:"/img/images1.jpeg",revision:"2b6dbda8b64caaf16e5c2635ea20f64b"},{url:"/img/logo.png",revision:"c5a5f06d2620af5980dbecebfcaa8747"},{url:"/manifest.json",revision:"b0b61d510feefab6ea8bdd50c2432a29"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:a,event:n,state:s})=>a&&"opaqueredirect"===a.type?new Response(a.body,{status:200,statusText:"OK",headers:a.headers}):a}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;const a=e.pathname;return!a.startsWith("/api/auth/")&&!!a.startsWith("/api/")},new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")},new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>!(self.origin===e.origin),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")});
