+++ src/app/about/components/Experience.tsx
@@ -4,124 +4,13 @@ import { motion } from "framer-motion"
 import {
   FaCalendarAlt,
   FaMapMarkerAlt,
-  FaLaptopCode, FaUniversity, FaBriefcase, FaGraduationCap, FaLaptop, FaCode
 } from 'react-icons/fa'
-
-const experienceAndStudies = [
-  {
-    id: '1',
-    title: 'Technical Analyst and Software Developer',
-    icon: FaUniversity,
-    company: 'CENSA (Centro de Sistemas de Antioquia)',
-    location: 'Medellín, Colombia',
-    description:
-      'Training with a focus on algorithms, data structures, and agile methodologies. Proficiency in languages such as JavaScript, Java, C#, HTML, and databases like MySQL and MongoDB.',
-    date: '2023 - 2024',
-    type: 'education',
-    skills: ['JavaScript', 'Java', 'C#', 'HTML', 'MySQL', 'MongoDB'],
-  },
-  {
-    id: '2',
-    title: 'Freelance Developer',
-    icon: FaBriefcase,
-    company: 'Independent Projects',
-    location: 'Remote',
-    description:
-      'Freelance experience in personalized projects, using technologies such as HTML, CSS, JavaScript, React, Node.js, NestJS, Next.js, SQL, and MongoDB.',
-    date: 'April 2023',
-    type: 'work',
-    skills: ['HTML', 'CSS', 'JavaScript', 'React', 'Node.js', 'NestJS', 'Next.js', 'SQL', 'MongoDB'],
-  },
-  {
-    id: '3',
-    title: 'Web Developer',
-    icon: FaLaptop,
-    company: 'Udemy',
-    location: 'Online',
-    description:
-      'Intensive course on complete web development, covering front-end technologies like HTML, CSS, JavaScript, and React.',
-    date: '2023',
-    type: 'education',
-    skills: ['HTML', 'CSS', 'JavaScript', 'React'],
-  },
-  {
-    id: '4',
-    title: 'Software Developer',
-    icon: FaLaptopCode,
-    company: 'BPO Global Services',
-    location: 'Medellín, Colombia',
-    description:
-      'Development and maintenance of web applications using front-end and back-end technologies such as Angular, Node.js, MySQL, SQL Server, CSS, Next.js, and Express.',
-    date: '2024 - 2025',
-    type: 'work',
-    skills: ['Angular', 'Node.js', 'MySQL', 'SQL Server', 'CSS', 'Next.js', 'Express'],
-  },
-  {
-    id: '5',
-    title: 'Full Stack Developer',
-    icon: FaLaptopCode,
-    company: 'Riwi.io',
-    location: 'Medellín, Colombia',
-    description:
-      'Learning contract developing web applications and custom software with NestJS, TypeScript, Node.js, MongoDB, SQL, Next.js, React, HTML, CSS. Participated in API integrations (Firebase Auth, Moodle) and developed soft skills.',
-    date: '2023 - 2024',
-    type: 'work',
-    skills: ['NestJS', 'TypeScript', 'Node.js', 'MongoDB', 'SQL', 'Next.js', 'React', 'HTML', 'CSS', 'Firebase', 'Moodle'],
-  },
-  {
-    id: '6',
-    title: 'Python Testing',
-    icon: FaLaptopCode,
-    company: 'CENSA (Centro de Sistemas de Antioquia)',
-    location: 'Medellín, Colombia',
-    description:
-      'Course on Python and unit testing using unittest, Selenium, and Faker for testing forms and web applications.',
-    date: '2024 - 2025',
-    type: 'education',
-    skills: ['Python', 'unittest', 'Selenium', 'Faker'],
-  },
-  {
-    id: '7',
-    title: 'Methodologies and Competencies',
-    icon: FaGraduationCap,
-    company: 'Various (YouTube, MOOCs)',
-    location: 'Online',
-    description:
-      'Courses about agile methodologies like Scrum and tools such as Jira, Confluence, and ClickUp, focused on leadership and managing high-performance teams.',
-    date: '2024 - 2025',
-    type: 'education',
-    skills: ['Scrum', 'Jira', 'Confluence', 'ClickUp', 'Leadership'],
-  },
-  {
-    id: '8',
-    title: 'Git and GitHub',
-    icon: FaCode,
-    company: 'Self-taught / Practical Projects',
-    location: 'Remote',
-    description:
-      'Proficiency in version control with Git and GitHub, including Git Flow and branching strategies in web and custom software projects.',
-    date: '2023 - 2024',
-    type: 'education',
-    skills: ['Git', 'GitHub', 'Git Flow', 'Branching Strategies'],
-  },
-  {
-    id: '9',
-    title: 'Learning Plan 2025',
-    icon: FaGraduationCap,
-    company: 'Personal Development Plan',
-    location: 'Remote',
-    description:
-      'For the year 2025, I plan to continue learning and strengthening my technical and soft skills, focusing on new technologies and agile methodologies.',
-    date: '2025 - 2026',
-    type: 'education',
-    skills: ['Continuous Learning', 'Agile', 'Leadership'],
-  },
-];
+import { experienceAndStudies } from "./experienceAndEstudies"
 
 
 export default function ExperienceAndEducation() {
   return (
-    <section className="py-24 relative overflow-hidden">
+    <section className="py-20 relative overflow-hidden">
       {/* Background Effects */}
       <div className="absolute inset-0 bg-gradient-to-b from-transparent via-purple-900/5 to-transparent" />
 
@@ -132,7 +21,7 @@ export default function ExperienceAndEducation() {
           animate={{ opacity: 1, y: 0 }}
           transition={{ duration: 0.8 }}
         >
-          <h2 className="text-5xl font-bold mb-6">
+          <h2 className="text-4xl font-bold mb-6">
             <span className="bg-clip-text text-transparent bg-gradient-to-r from-purple-400 via-pink-500 to-purple-600">
               Experience & Education
             </span>
@@ -157,7 +46,7 @@ export default function ExperienceAndEducation() {
                 transition={{ duration: 0.6, delay: index * 0.2 }}
               >
                 {/* Content Card */}
-                <div className={`w-full lg:w-5/12 ${index % 2 === 0 ? "lg:text-right" : "lg:text-left"}`}>
+                <div className={`w-full lg:w-5/12 ${index % 2 === 0 ? "lg:text-start" : "lg:text-left"}`}>
                   <motion.div
                     className="group relative bg-gray-900/50 backdrop-blur-sm rounded-3xl p-8 border border-gray-800/50 hover:border-purple-500/50 transition-all duration-500"
                     whileHover={{ scale: 1.02, y: -5 }}
@@ -249,8 +138,8 @@ export default function ExperienceAndEducation() {
           transition={{ duration: 0.8, delay: 1 }}
         >
           {[
-            { number: "3+", label: "Years Experience", color: "from-blue-500 to-cyan-500" },
-            { number: "15+", label: "Projects Completed", color: "from-purple-500 to-pink-500" },
+            { number: "2+", label: "Years Experience", color: "from-blue-500 to-cyan-500" },
+            { number: "10+", label: "Projects Completed", color: "from-purple-500 to-pink-500" },
             { number: "5+", label: "Technologies", color: "from-green-500 to-emerald-500" },
             { number: "100%", label: "Client Satisfaction", color: "from-orange-500 to-red-500" },
           ].map((stat, index) => (
