+++ src/app/about/components/experienceAndEstudies.ts
@@ -1,0 +1,114 @@
+import { FaBriefcase, FaCode, FaGraduationCap, FaLaptop, FaLaptopCode, FaUniversity } from "react-icons/fa";
+
+
+export const experienceAndStudies = [
+  {
+    id: '1',
+    title: 'Technical Analyst and Software Developer',
+    icon: FaUniversity,
+    company: 'CENSA (Centro de Sistemas de Antioquia)',
+    location: 'Medellín, Colombia',
+    description:
+      'Training with a focus on algorithms, data structures, and agile methodologies. Proficiency in languages such as JavaScript, Java, C#, HTML, and databases like MySQL and MongoDB.',
+    date: '2023 - 2024',
+    type: 'education',
+    skills: ['JavaScript', 'Java', 'C#', 'HTML', 'MySQL', 'MongoDB'],
+  },
+  {
+    id: '2',
+    title: 'Freelance Developer',
+    icon: FaBriefcase,
+    company: 'Independent Projects',
+    location: 'Remote',
+    description:
+      'Freelance experience in personalized projects, using technologies such as HTML, CSS, JavaScript, React, Node.js, NestJS, Next.js, SQL, and MongoDB.',
+    date: 'April 2023',
+    type: 'work',
+    skills: ['HTML', 'CSS', 'JavaScript', 'React', 'Node.js', 'NestJS', 'Next.js', 'SQL', 'MongoDB'],
+  },
+  {
+    id: '3',
+    title: 'Web Developer',
+    icon: FaLaptop,
+    company: 'Udemy',
+    location: 'Online',
+    description:
+      'Intensive course on complete web development, covering front-end technologies like HTML, CSS, JavaScript, and React.',
+    date: '2023',
+    type: 'education',
+    skills: ['HTML', 'CSS', 'JavaScript', 'React'],
+  },
+  {
+    id: '4',
+    title: 'Software Developer',
+    icon: FaLaptopCode,
+    company: 'BPO Global Services',
+    location: 'Medellín, Colombia',
+    description:
+      'Development and maintenance of web applications using front-end and back-end technologies such as Angular, Node.js, MySQL, SQL Server, CSS, Next.js, and Express.',
+    date: '2024 - 2025',
+    type: 'work',
+    skills: ['Angular', 'Node.js', 'MySQL', 'SQL Server', 'CSS', 'Next.js', 'Express'],
+  },
+  {
+    id: '5',
+    title: 'Full Stack Developer',
+    icon: FaLaptopCode,
+    company: 'Riwi.io',
+    location: 'Medellín, Colombia',
+    description:
+      'Learning contract developing web applications and custom software with NestJS, TypeScript, Node.js, MongoDB, SQL, Next.js, React, HTML, CSS. Participated in API integrations (Firebase Auth, Moodle) and developed soft skills.',
+    date: '2023 - 2024',
+    type: 'work',
+    skills: ['NestJS', 'TypeScript', 'Node.js', 'MongoDB', 'SQL', 'Next.js', 'React', 'HTML', 'CSS', 'Firebase', 'Moodle'],
+  },
+  {
+    id: '6',
+    title: 'Python Testing',
+    icon: FaLaptopCode,
+    company: 'CENSA (Centro de Sistemas de Antioquia)',
+    location: 'Medellín, Colombia',
+    description:
+      'Course on Python and unit testing using unittest, Selenium, and Faker for testing forms and web applications.',
+    date: '2024 - 2025',
+    type: 'education',
+    skills: ['Python', 'unittest', 'Selenium', 'Faker'],
+  },
+  {
+    id: '7',
+    title: 'Methodologies and Competencies',
+    icon: FaGraduationCap,
+    company: 'Various (YouTube, MOOCs)',
+    location: 'Online',
+    description:
+      'Courses about agile methodologies like Scrum and tools such as Jira, Confluence, and ClickUp, focused on leadership and managing high-performance teams.',
+    date: '2024 - 2025',
+    type: 'education',
+    skills: ['Scrum', 'Jira', 'Confluence', 'ClickUp', 'Leadership'],
+  },
+  {
+    id: '8',
+    title: 'Git and GitHub',
+    icon: FaCode,
+    company: 'Self-taught / Practical Projects',
+    location: 'Remote',
+    description:
+      'Proficiency in version control with Git and GitHub, including Git Flow and branching strategies in web and custom software projects.',
+    date: '2023 - 2024',
+    type: 'education',
+    skills: ['Git', 'GitHub', 'Git Flow', 'Branching Strategies'],
+  },
+  {
+    id: '9',
+    title: 'Learning Plan 2025',
+    icon: FaGraduationCap,
+    company: 'Personal Development Plan',
+    location: 'Remote',
+    description:
+      'For the year 2025, I plan to continue learning and strengthening my technical and soft skills, focusing on new technologies and agile methodologies.',
+    date: '2025 - 2026',
+    type: 'education',
+    skills: ['Continuous Learning', 'Agile', 'Leadership'],
+  },
+];
+
