+++ src/app/api/email/route.ts
@@ -2,7 +2,7 @@ import { NextRequest, NextResponse } from 'next/server'
 import nodemailer from 'nodemailer'
 
 export async function POST(req: NextRequest) {
-  const { email, asunto, mensaje } = await req.json()
+  const { name, email, subject, message } = await req.json()
 
   // Configuración del transporte de Nodemailer
   const transporter = nodemailer.createTransport({
@@ -18,39 +18,38 @@ export async function POST(req: NextRequest) {
     <!DOCTYPE html>
     <html lang="es">
     <head>
-      <meta charset="UTF-8">
-      <meta name="viewport" content="width=device-width, initial-scale=1.0">
+      <meta charset="UTF-8" />
+      <meta name="viewport" content="width=device-width, initial-scale=1.0" />
       <title>Nuevo mensaje de contacto</title>
       <style>
         @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;600&display=swap');
         body {
-          font-family: 'Poppins', Arial, sans-serif;
-          line-height: 1.6;
-          color: #f8fafc;
-          background-color: #0f172a;
           margin: 0;
           padding: 0;
+          font-family: 'Poppins', sans-serif;
+          background-color: #0f172a;
+          color: #f8fafc;
         }
         .container {
-          max-width: 600px;
-          margin: 20px auto;
-          padding: 20px;
+          max-width: 640px;
+          margin: 30px auto;
+          padding: 24px;
         }
         .card {
           background-color: #1e293b;
-          border-radius: 12px;
-          box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05);
+          border-radius: 16px;
           overflow: hidden;
+          box-shadow: 0 20px 30px rgba(0, 0, 0, 0.25);
+          border: 1px solid #334155;
         }
         .header {
-          background: linear-gradient(90deg, #a78bfa, #ec4899);
-          color: white;
-          padding: 20px;
-          font-size: 24px;
-          font-weight: bold;
+          background: linear-gradient(to right, #a78bfa, #ec4899);
+          padding: 24px;
+          color: #ffffff;
           text-align: center;
-          text-transform: uppercase;
-          letter-spacing: 2px;
+          font-size: 20px;
+          font-weight: 600;
+          letter-spacing: 1px;
         }
         .content {
           padding: 30px;
@@ -59,32 +58,31 @@ export async function POST(req: NextRequest) {
           margin-bottom: 20px;
         }
         .field-label {
-          display: block;
           font-weight: 600;
-          margin-bottom: 5px;
+          margin-bottom: 6px;
+          display: block;
           color: #a78bfa;
+          font-size: 14px;
         }
         .field-value {
           background-color: #2d3748;
-          padding: 10px;
-          border-radius: 6px;
+          padding: 12px 16px;
+          border-radius: 8px;
+          font-size: 14px;
           color: #f8fafc;
-        }
-        .message {
           white-space: pre-wrap;
         }
         .footer {
+          margin-top: 30px;
           text-align: center;
-          margin-top: 20px;
           font-size: 12px;
-          color: #cbd5e0;
+          color: #94a3b8;
         }
-        .icon {
-          display: inline-block;
+        svg.icon {
           width: 20px;
           height: 20px;
-          margin-right: 10px;
           vertical-align: middle;
+          margin-right: 8px;
         }
       </style>
     </head>
@@ -92,7 +90,7 @@ export async function POST(req: NextRequest) {
       <div class="container">
         <div class="card">
           <div class="header">
-            <svg class="icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor">
+            <svg class="icon" fill="currentColor" viewBox="0 0 24 24">
               <path d="M1.5 8.67v8.58a3 3 0 003 3h15a3 3 0 003-3V8.67l-8.928 5.493a3 3 0 01-3.144 0L1.5 8.67z" />
               <path d="M22.5 6.908V6.75a3 3 0 00-3-3h-15a3 3 0 00-3 3v.158l9.714 5.978a1.5 1.5 0 001.572 0L22.5 6.908z" />
             </svg>
@@ -100,16 +98,20 @@ export async function POST(req: NextRequest) {
           </div>
           <div class="content">
             <div class="field">
-              <span class="field-label">De:</span>
+              <span class="field-label">Nombre:</span>
+              <div class="field-value">${name}</div>
+            </div>
+            <div class="field">
+              <span class="field-label">Correo Electrónico:</span>
               <div class="field-value">${email}</div>
             </div>
             <div class="field">
               <span class="field-label">Asunto:</span>
-              <div class="field-value">${asunto}</div>
+              <div class="field-value">${subject}</div>
             </div>
             <div class="field">
               <span class="field-label">Mensaje:</span>
-              <div class="field-value message">${mensaje}</div>
+              <div class="field-value">${message}</div>
             </div>
           </div>
         </div>
@@ -119,12 +121,12 @@ export async function POST(req: NextRequest) {
       </div>
     </body>
     </html>
-  `
+  `;
 
   const mailOptions = {
     from: email,
     to: process.env.RECEIVER_EMAIL,
-    subject: asunto,
+    subject,
     html: htmlTemplate,
     replyTo: email,
   }
