+++ src/app/layout.tsx
@@ -8,6 +8,7 @@ import Footer from "./components/Footer"
 const inter = Inter({ subsets: ["latin"] })
 
 export const metadata: Metadata = {
+  metadataBase: new URL("https://cristdev.vercel.app"),
   title: {
     default: "Cristian Manco | Full Stack Developer",
     template: "%s | Cristian Manco",
@@ -39,7 +40,7 @@ export const metadata: Metadata = {
     siteName: "Cristian Manco Portfolio",
     images: [
       {
-        url: "/og-image.jpg",
+        url: "/img/logo.png",
         width: 1200,
         height: 630,
         alt: "Cristian Manco - Full Stack Developer",
@@ -51,7 +52,7 @@ export const metadata: Metadata = {
     title: "Cristian Manco | Full Stack Developer",
     description:
       "Passionate Full Stack Developer creating innovative digital solutions. Specializing in React, Node.js, and modern web technologies.",
-    images: ["/og-image.jpg"],
+    images: ["/img/logo.png"],
     creator: "@cristiandev",
   },
   robots: {
@@ -66,7 +67,7 @@ export const metadata: Metadata = {
     },
   },
   verification: {
-    google: "your-google-verification-code",
+    google: "google-verification-code",
   },
 }
 
@@ -78,12 +79,17 @@ export default function RootLayout({
   return (
     <html lang="en" className="scroll-smooth">
       <head>
+        {/* PWA */}
+        <link rel="manifest" href="/manifest.json" />
+        <meta name="theme-color" content="#8b5cf6" />
+        <meta name="apple-mobile-web-app-capable" content="yes" />
+        <meta name="mobile-web-app-capable" content="yes" />
+
+        {/* Favicons */}
         <link rel="icon" href="/favicon.ico" />
         <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
-        <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
-        <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
-        <link rel="manifest" href="/site.webmanifest" />
-        <meta name="theme-color" content="#8b5cf6" />
+        <link rel="icon" type="image/png" sizes="32x32" href="/icons/icon-32x32.png" />
+        <link rel="icon" type="image/png" sizes="16x16" href="/icons/icon-16x16.png" />
       </head>
       <body className={`${inter.className} antialiased`}>
         <Header />
