+++ src/app/skills/components/professionalSkill.ts
@@ -1,0 +1,75 @@
+import { FaClock, FaComments, FaHandshake, FaHeart, FaLightbulb, FaProjectDiagram, FaShieldAlt, FaSync, FaUsers } from "react-icons/fa";
+
+
+export const professionalSkills = [
+  {
+    name: "Communication",
+    level: 90,
+    icon: FaComments,
+    color: "from-blue-500 to-cyan-500",
+    description: "Clear and effective communication",
+  },
+  {
+    name: "Teamwork",
+    level: 97,
+    icon: FaUsers,
+    color: "from-purple-500 to-pink-500",
+    description: "Collaborative problem solving",
+  },
+  {
+    name: "Creativity",
+    level: 95,
+    icon: FaLightbulb,
+    color: "from-yellow-500 to-orange-500",
+    description: "Innovative solutions and ideas",
+  },
+  {
+    name: "Discipline",
+    level: 90,
+    icon: FaShieldAlt,
+    color: "from-gray-500 to-gray-600",
+    description: "Consistent work ethic",
+  },
+  {
+    name: "Dedication",
+    level: 98,
+    icon: FaHeart,
+    color: "from-red-500 to-pink-500",
+    description: "Passionate about development",
+  },
+  {
+    name: "Responsibility",
+    level: 99,
+    icon: FaShieldAlt,
+    color: "from-green-500 to-emerald-500",
+    description: "Accountable and reliable",
+  },
+  {
+    name: "Project Management",
+    level: 94,
+    icon: FaProjectDiagram,
+    color: "from-indigo-500 to-purple-500",
+    description: "Agile methodologies, planning",
+  },
+  {
+    name: "Time Management",
+    level: 96,
+    icon: FaClock,
+    color: "from-orange-500 to-red-500",
+    description: "Efficient task prioritization",
+  },
+  {
+    name: "Adaptation to Change",
+    level: 99,
+    icon: FaSync,
+    color: "from-teal-500 to-cyan-500",
+    description: "Flexible and adaptable",
+  },
+  {
+    name: "Empathy",
+    level: 97,
+    icon: FaHandshake,
+    color: "from-pink-500 to-rose-500",
+    description: "Understanding team dynamics",
+  },
+]
