+++ src/app/skills/page.tsx
@@ -1,4 +1,4 @@
-import Skills from "@/app/components/Skills";
+import Skills from "@/app/skills/components/Skills";
 
 export default function SkillsPage() {
   return <Skills />
